Rename typing.py
It conflicts with python's own typing module.
