+++ blueman/gui/GtkAnimation.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Collection, Iterable
 import cairo
 import gi
 
-from blueman.typing import GSignals
+from blueman.bluemantyping import GSignals
 
 gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk
