+++ blueman/gui/manager/ManagerProgressbar.py
@@ -4,7 +4,7 @@ from typing import List, TYPE_CHECKING, Callable
 
 import gi
 
-from blueman.typing import GSignals
+from blueman.bluemantyping import GSignals
 
 gi.require_version("Gtk", "3.0")
 gi.require_version("Gdk", "3.0")
