+++ blueman/plugins/applet/NetUsage.py
@@ -17,7 +17,7 @@ from gi.repository import GLib
 import gi
 
 from blueman.plugins.applet.PPPSupport import PPPConnectedListener
-from blueman.typing import GSignals
+from blueman.bluemantyping import GSignals
 
 gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk
